+++ packages/synapse-interface/utils/types/index.tsx
@@ -70,6 +70,9 @@ type QuoteQuery = {
 }
 
 export type BridgeQuote = {
+  inputAmountForQuote: string
+  originTokenForQuote: Token
+  destTokenForQuote: Token
   outputAmount: bigint
   outputAmountString: string
   routerAddress: string
@@ -85,6 +88,7 @@ export type BridgeQuote = {
   timestamp: number
   originChainId: number
   destChainId: number
+  requestId: number
 }
 
 interface TokensByChain {