+++ packages/synapse-interface/CHANGELOG.md
@@ -3,6 +3,109 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.36.1](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.36.0...@synapsecns/synapse-interface@0.36.1) (2024-08-26)
+
+**Note:** Version bump only for package @synapsecns/synapse-interface
+
+
+
+
+
+# [0.36.0](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.35.1...@synapsecns/synapse-interface@0.36.0) (2024-08-26)
+
+
+### Features
+
+* **synapse-interface:** bridge quote state validations ([#3019](https://github.com/synapsecns/sanguine/issues/3019)) ([649b3f5](https://github.com/synapsecns/sanguine/commit/649b3f597787a3f81d98fdf64adf81bad851a45d)), closes [#3018](https://github.com/synapsecns/sanguine/issues/3018)
+
+
+
+
+
+## [0.35.1](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.35.0...@synapsecns/synapse-interface@0.35.1) (2024-08-25)
+
+**Note:** Version bump only for package @synapsecns/synapse-interface
+
+
+
+
+
+# [0.35.0](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.34.7...@synapsecns/synapse-interface@0.35.0) (2024-08-21)
+
+
+### Features
+
+* **synapse-interface:** METIS token ([#3044](https://github.com/synapsecns/sanguine/issues/3044)) ([d73373b](https://github.com/synapsecns/sanguine/commit/d73373b93395fbe09064671c39ec774ca521330f))
+
+
+
+
+
+## [0.34.7](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.34.6...@synapsecns/synapse-interface@0.34.7) (2024-08-19)
+
+**Note:** Version bump only for package @synapsecns/synapse-interface
+
+
+
+
+
+## [0.34.6](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.34.5...@synapsecns/synapse-interface@0.34.6) (2024-08-16)
+
+
+### Bug Fixes
+
+* **synapse-interface:** hide Blast WETH ([#3033](https://github.com/synapsecns/sanguine/issues/3033)) ([8da405f](https://github.com/synapsecns/sanguine/commit/8da405f81183817f6564e859914d2dc02a49735f))
+
+
+
+
+
+## [0.34.5](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.34.4...@synapsecns/synapse-interface@0.34.5) (2024-08-15)
+
+**Note:** Version bump only for package @synapsecns/synapse-interface
+
+
+
+
+
+## [0.34.4](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.34.3...@synapsecns/synapse-interface@0.34.4) (2024-08-14)
+
+
+### Bug Fixes
+
+* **synapse-interface:** don't adjust `payload.value` that SDK already handles ([#3015](https://github.com/synapsecns/sanguine/issues/3015)) ([77f5863](https://github.com/synapsecns/sanguine/commit/77f58637d948c981b9e26f51b2137fd2f1e0578f))
+
+
+
+
+
+## [0.34.3](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.34.2...@synapsecns/synapse-interface@0.34.3) (2024-08-13)
+
+
+### Bug Fixes
+
+* **synapse-interface:** disable bridge ui pending wallet ([#3014](https://github.com/synapsecns/sanguine/issues/3014)) ([fbe4b60](https://github.com/synapsecns/sanguine/commit/fbe4b60b952b52b8f5cbec01eae8a35f100caf81))
+
+
+
+
+
+## [0.34.2](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.34.1...@synapsecns/synapse-interface@0.34.2) (2024-08-13)
+
+**Note:** Version bump only for package @synapsecns/synapse-interface
+
+
+
+
+
+## [0.34.1](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.34.0...@synapsecns/synapse-interface@0.34.1) (2024-08-13)
+
+**Note:** Version bump only for package @synapsecns/synapse-interface
+
+
+
+
+
 # [0.34.0](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.33.0...@synapsecns/synapse-interface@0.34.0) (2024-08-09)